requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
auto_gptq==0.7.1
bcrypt==4.0.1
beautifulsoup4==4.13.4
bitsandbytes==0.45.5
cachetools==5.5.2
cbor==1.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.6.0
dill==0.3.8
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
faiss-cpu==1.8.0
fastapi==0.115.4
filelock==3.18.0
FlagEmbedding==1.3.4
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.0
gekko==1.3.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
greenlet==3.2.2
gritql==0.2.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
humanfriendly==10.0
idna==3.10
ijson==3.4.0
inscriptis==2.6.0
ir_datasets==0.5.10
Jinja2==3.1.6
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.22
langchain-cli==0.0.36
langchain-community==0.3.7
langchain-core==0.3.59
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.8
langserve==0.3.1
langsmith==0.1.147
lxml==5.4.0
lz4==4.4.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.0
mdurl==0.1.2
motor==3.6.0
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
optimum==1.17.0
orjson==3.10.18
packaging==24.2
pandas==2.2.2
passlib==1.7.4
peft==0.15.2
pillow==11.1.0
plotly==5.22.0
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
pyarrow==20.0.0
pyasn1
pyasn1_modules
pydantic==2.11.1
pydantic-settings==2.9.1
pydantic_core==2.33.0
Pygments==2.19.1
PyJWT==2.9.0
pymongo==4.9.1
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.12
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rouge==1.0.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.0.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.35
sse-starlette==1.8.2
starlette==0.41.0
sympy==1.13.3
ta==0.11.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
tomlkit==0.13.2
tqdm==4.67.1
transformers==4.51.3
trec-car-tools==2.6
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
unlzw3==0.2.3
urllib3==2.4.0
uvicorn==0.34.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchfiles==1.0.5
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
zlib-state==0.1.9
